+++ src/@core/runner/runner.types.ts
@@ -5,8 +5,12 @@ export type ProjectContext = WorkflowContext
 export type Block = PipelineStep
 
 export interface Hooks {
-  projectStart(context: WorkflowContext): Promise<void>
-  projectEnd(context: WorkflowContext): Promise<void>
-  blockStart(context: ProjectContext, block: Block): Promise<void>
-  blockEnd(context: ProjectContext, block: Block): Promise<void>
+  // advanced lifecycle hooks
+  onMessage?(event: any, context: ProjectContext): Promise<void>
+
+  // project/block hooks
+  projectStart?(context: ProjectContext): Promise<void>
+  projectEnd?(context: ProjectContext): Promise<void>
+  blockStart?(context: ProjectContext, block: Block): Promise<void>
+  blockEnd?(context: ProjectContext, block: Block): Promise<void>
 }